Moved config.go to prevent conflicts with monorepo
