+++ cmd/reader/main.go
@@ -7,9 +7,9 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/seventv/7tv-bot/config"
 	"github.com/seventv/7tv-bot/internal/database"
 	"github.com/seventv/7tv-bot/internal/irc-reader"
+	"github.com/seventv/7tv-bot/internal/irc-reader/config"
 )
 
 func main() {
