+++ internal/irc-reader/config/config.go
